Update WipeoutAll to prevent incorrect shredding.
In summary, using Shredder to shred some data in RT would
sometimes delete additional or incorrect data. One bug report
available here:

http://issues.bestpractical.com/Ticket/Display.html?id=16070

The author tracked down the source of the problem and commented
on the ticket:

"The problem seems to lie in the WipeoutAll function of Shredder.pm.

If you replace with the code from RTx::Shredder 0.07, wiping out users
behaves normally again, and no tickets are improperly shredded."

The root of the issue is using the each iterator with
deletes.

This commit includes the change and tests.

In the process of writing the tests, it was discovered that the result
set from dump_current_and_savepoint was empty. Setting the schema
arg in table_info fixes this and led to the following change:

Modify DB check in ticket test to reflect conditions after WipeoutAll.

The commit modifying the table_info call in utils
(27b8a617) revealed the DBs in shredder tests appeared to be empty
and one test in 01ticket.t started failing. This commit removes
a link and a transaction for the parent that are still
in the DB after a child ticket is wiped. Need to confirm this
is the expected behavior. These are successfully removed by
shredder when the parent is wiped.